+++ djstripe/models/connect.py
@@ -1,3 +1,5 @@
+import warnings
+
 import stripe
 from django.db import models
 
@@ -17,14 +19,31 @@ class Account(StripeModel):
 	"""
 
 	stripe_class = stripe.Account
-
-	business_logo = models.ForeignKey("FileUpload", on_delete=models.SET_NULL, null=True)
+	# Special handling of the icon and logo fields, they moved to settings.branding in Stripe 2019-02-19 but we
+	# want them as ForeignKeys
+	branding_icon = models.ForeignKey(
+		"FileUpload",
+		on_delete=models.SET_NULL,
+		null=True,
+		related_name="icon_account",
+		help_text="An icon for the account. Must be square and at least 128px x 128px.",
+	)
+	branding_logo = models.ForeignKey(
+		"FileUpload",
+		on_delete=models.SET_NULL,
+		null=True,
+		related_name="logo_account",
+		help_text="A logo for the account that will be used in Checkout instead of the icon "
+		"and without the account’s name next to it if provided. Must be at least 128px x 128px.",
+	)
+	# deprecated, will be removed in 2.2.  see https://stripe.com/docs/upgrades#2019-02-19
 	business_name = models.CharField(
 		max_length=255,
 		default="",
 		blank=True,
 		help_text="The publicly visible name of the business",
 	)
+	# deprecated, will be removed in 2.2. see https://stripe.com/docs/upgrades#2019-02-19
 	business_primary_color = models.CharField(
 		max_length=7,
 		default="",
@@ -33,6 +52,13 @@ class Account(StripeModel):
 			"A CSS hex color value representing the primary branding color for this account"
 		),
 	)
+	business_profile = JSONField(
+		null=True, blank=True, help_text=("Optional information related to the business.")
+	)
+	business_type = StripeEnumField(
+		enum=enums.BusinessType, default="", blank=True, help_text="The business type."
+	)
+	# deprecated, will be removed in 2.2. see https://stripe.com/docs/upgrades#2019-02-19
 	business_url = models.CharField(
 		max_length=200,
 		default="",
@@ -43,6 +69,15 @@ class Account(StripeModel):
 		help_text="Whether the account can create live charges"
 	)
 	country = models.CharField(max_length=2, help_text="The country of the account")
+	company = JSONField(
+		null=True,
+		blank=True,
+		help_text=(
+			"Information about the company or business. "
+			"This field is null unless business_type is set to company."
+		),
+	)
+	# deprecated, will be removed in 2.2. see https://stripe.com/docs/upgrades#2019-02-19
 	debit_negative_balances = models.NullBooleanField(
 		null=True,
 		blank=True,
@@ -52,6 +87,7 @@ class Account(StripeModel):
 			"balances from an attached bank account."
 		),
 	)
+	# deprecated, will be removed in 2.2. see https://stripe.com/docs/upgrades#2019-02-19
 	decline_charge_on = JSONField(
 		null=True,
 		blank=True,
@@ -69,6 +105,7 @@ class Account(StripeModel):
 			"Standard accounts cannot receive payouts before this is true."
 		)
 	)
+	# deprecated, will be removed in 2.2. see https://stripe.com/docs/upgrades#2019-02-19
 	display_name = models.CharField(
 		max_length=255,
 		default="",
@@ -80,6 +117,15 @@ class Account(StripeModel):
 	)
 	email = models.CharField(max_length=255, help_text="The primary user’s email address.")
 	# TODO external_accounts = ...
+	individual = JSONField(
+		null=True,
+		blank=True,
+		help_text=(
+			"Information about the person represented by the account. "
+			"This field is null unless business_type is set to individual."
+		),
+	)
+	# deprecated, will be removed in 2.2. see https://stripe.com/docs/upgrades#2019-02-19
 	legal_entity = JSONField(
 		null=True,
 		blank=True,
@@ -87,6 +133,7 @@ class Account(StripeModel):
 			"Information about the legal entity itself, including about the associated account representative"
 		),
 	)
+	# deprecated, will be removed in 2.2. see https://stripe.com/docs/upgrades#2019-02-19
 	payout_schedule = JSONField(
 		null=True,
 		blank=True,
@@ -94,6 +141,7 @@ class Account(StripeModel):
 			"Details on when funds from charges are available, and when they are paid out to an external account."
 		),
 	)
+	# deprecated, will be removed in 2.2. see https://stripe.com/docs/upgrades#2019-02-19
 	payout_statement_descriptor = models.CharField(
 		max_length=255,
 		default="",
@@ -112,6 +160,22 @@ class Account(StripeModel):
 			"It’s used by Stripe for risk and underwriting purposes."
 		),
 	)
+	requirements = JSONField(
+		null=True,
+		blank=True,
+		help_text=(
+			"Information about the requirements for the account, "
+			"including what information needs to be collected, and by when."
+		),
+	)
+	settings = JSONField(
+		null=True,
+		blank=True,
+		help_text=(
+			"Account options for customizing how the account functions within Stripe."
+		),
+	)
+	# deprecated, will be removed in 2.2. see https://stripe.com/docs/upgrades#2019-02-19
 	statement_descriptor = models.CharField(
 		max_length=255,
 		default="",
@@ -120,24 +184,28 @@ class Account(StripeModel):
 			"The default text that appears on credit card statements when a charge is made directly on the account"
 		),
 	)
+	# deprecated, will be removed in 2.2. see https://stripe.com/docs/upgrades#2019-02-19
 	support_email = models.CharField(
 		max_length=255,
 		default="",
 		blank=True,
 		help_text="A publicly shareable support email address for the business",
 	)
+	# deprecated, will be removed in 2.2. see https://stripe.com/docs/upgrades#2019-02-19
 	support_phone = models.CharField(
 		max_length=255,
 		default="",
 		blank=True,
 		help_text="A publicly shareable support phone number for the business",
 	)
+	# deprecated, will be removed in 2.2. see https://stripe.com/docs/upgrades#2019-02-19
 	support_url = models.CharField(
 		max_length=200,
 		default="",
 		blank=True,
 		help_text="A publicly shareable URL that provides support for this account",
 	)
+	# deprecated, will be removed in 2.2. see https://stripe.com/docs/upgrades#2019-02-19
 	timezone = models.CharField(
 		max_length=50, help_text="The timezone used in the Stripe Dashboard for this account."
 	)
@@ -147,6 +215,7 @@ class Account(StripeModel):
 		blank=True,
 		help_text="Details on the acceptance of the Stripe Services Agreement",
 	)
+	# deprecated, will be removed in 2.2. see https://stripe.com/docs/upgrades#2019-02-19
 	verification = JSONField(
 		null=True,
 		blank=True,
@@ -170,6 +239,97 @@ class Account(StripeModel):
 
 	def __str__(self):
 		return self.display_name or self.business_name
+
+	# deprecated, remove in 2.2
+	@property
+	def business_logo(self):
+		warnings.warn(
+			"Account.business_logo has been renamed to branding_icon, this alias will be removed in dj-stripe 2.2",
+			DeprecationWarning,
+		)
+		return self.branding_icon
+
+	@classmethod  # noqa: C901
+	def _manipulate_stripe_object_hook(cls, data):
+		data = super()._manipulate_stripe_object_hook(data)
+
+		def empty_string_to_none(v):
+			"""
+			stripe.StripeObject.__setitem__ doesn't allow = ""
+			"""
+			if v == "":
+				return None
+			else:
+				return v
+
+		# icon (formerly called business_logo) logo (formerly called business_logo_large)
+		# moved to settings.branding in Stripe 2019-02-19 but we'll keep them to provide the ForeignKey
+		for old, new in [("branding_icon", "icon"), ("branding_logo", "logo")]:
+			try:
+				data[old] = data["settings"]["branding"][new]
+			except KeyError:
+				pass
+
+		# copy data back from new location to deprecated fields to be removed in 2.2
+		# see https://stripe.com/docs/upgrades#2019-02-19
+		try:
+			new = data["requirements"]["current_deadline"]
+			if new:
+				data["verification"] = data["verification"] or {}
+				data["verification"]["due_by"] = new
+		except KeyError:
+			pass
+
+		for old, new in [
+			("payout_schedule", "schedule"),
+			("payout_statement_descriptor", "statement_descriptor"),
+			("debit_negative_balances", "debit_negative_balances"),
+		]:
+			try:
+				data[old] = empty_string_to_none(data["settings"]["payouts"][new])
+			except KeyError:
+				pass
+
+		for old, new in [("statement_descriptor", "statement_descriptor")]:
+			try:
+				data[old] = empty_string_to_none(data["settings"]["payments"][new])
+			except KeyError:
+				pass
+
+		for old, new in [("decline_charge_on", "decline_on")]:
+			try:
+				data[old] = data["settings"]["card_payments"][new]
+			except KeyError:
+				pass
+
+		for old, new in [("business_primary_color", "primary_color")]:
+			try:
+				data[old] = empty_string_to_none(data["settings"]["branding"][new])
+			except KeyError:
+				pass
+
+		for old, new in [("display_name", "display_name"), ("timezone", "timezone")]:
+			try:
+				data[old] = empty_string_to_none(data["settings"]["dashboard"][new])
+			except KeyError:
+				pass
+
+		for old, new in [
+			("business_name", "name"),
+			("business_url", "url"),
+			("product_description", "product_description"),
+			("support_address", "support_address"),
+			("support_email", "support_email"),
+			("support_phone", "support_phone"),
+			("support_url", "support_url"),
+		]:
+			try:
+				data[old] = empty_string_to_none(data["business_profile"][new])
+			except KeyError:
+				pass
+		# end of deprecated fields to be removed in 2.2
+
+		return data
 
 
 class ApplicationFee(StripeModel):